feat: implement Authentication with jwt, login, register, and protected routes

+++ validator/book_validator.go
@@ -2,14 +2,14 @@ package validator
 
 type BookPostRequest struct {
 	Title       string `json:"title" binding:"required"`
-	Price       int    `json:"price" binding:"required,number"`
+	Price       int64  `json:"price" binding:"required,number"`
 	Description string `json:"description"`
 	Rating      int    `json:"rating" binding:"required,number"`
 }
 
 type BookUpdateRequest struct {
 	Title       string `json:"title" binding:"required"`
-	Price       int    `json:"price" binding:"required,number"`
+	Price       int64  `json:"price" binding:"required,number"`
 	Description string `json:"description"`
 	Rating      int    `json:"rating" binding:"required,number"`
 }
